Moved Wallet Connection logic into App to keep central for accessing from Profile, Nav, and Mint. Refactored Navabar.
