+++ eth-app/packages/react-app/src/App.js
@@ -1,9 +1,11 @@
-import React from "react";
+import React, { useEffect, useState } from "react";
+
+import useWeb3Modal from "./hooks/useWeb3Modal";
+
+
 // import { useQuery } from "@apollo/react-hooks";
 //import { Contract } from "@ethersproject/contracts";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
-// import useSessionStorage from "./hooks/useSessionStorage";
-// import useWeb3Modal from "./hooks/useWeb3Modal";
 
 
 //import { getDefaultProvider } from "@ethersproject/providers";
@@ -34,8 +36,40 @@ async function readOnChainData() {
 
 function App() {
 
+  const [account, setAccount] = useState("");
+  const [rendered, setRendered] = useState("");
+  const [provider, loadWeb3Modal, logoutOfWeb3Modal] = useWeb3Modal();
+
+  useEffect(() => {
+    async function fetchAccount() {
+      try {
+        if (!provider) {
+          return;
+        }
+
+        // Load the user's accounts.
+        const accounts = await provider.listAccounts();
+        setAccount(accounts[0]);
+
+        // Resolve the ENS name for the first account.
+        const name = await provider.lookupAddress(accounts[0]);
+
+        // Render either the ENS name or the shortened account address.
+        if (name) {
+          setRendered(name);
+        } else {
+          setRendered(account.substring(0, 6) + "..." + account.substring(36));
+        }
+      } catch (err) {
+        setAccount("");
+        setRendered("");
+        console.error(err);
+      }
+    }
+    fetchAccount();
+
+  }, [account, provider, setAccount, setRendered]);
   // const { loading, error, data } = useQuery(GET_TRANSFERS); // For use with GraphQL
-  // const [provider, loadWeb3Modal, logoutOfWeb3Modal] = useWeb3Modal();
   // For GraphQl
   // React.useEffect(() => {
   //   if (!loading && !error && data && data.transfers) {
@@ -43,11 +77,18 @@ function App() {
   //   }
   // }, [loading, error, data]);
 
+  function wallentButtonClicked() {
+    if (!provider) {
+      loadWeb3Modal();
+    } else {
+      logoutOfWeb3Modal();
+    }
+  }
 
   return (
     <Router>
       <div className="App">
-        <Navbar/>
+        <Navbar handleWalletConnect={wallentButtonClicked} rendered={!rendered ? "Connect Wallet" : rendered + ""}></Navbar>
         <Switch>
           <Route exact path="/">
             <Home />
